requirements.txt
black
click
einops
hydra-core
matplotlib
numpy
wandb
pytest
python-dotenv
torch
requests
tqdm
